Progress

+++ src/resolvers/Mutation.js
@@ -1,50 +1,57 @@
+import bcrypt from 'bcryptjs';
+
+import getUserId from '../utils/getUserId';
+import generateToken from '../utils/generateToken';
+import hashPassword from '../utils/hashPassword';
+
 const Mutation = {
   async createUser(parent, args, { prisma }, info) {
-    const emailTaken = await prisma.exists.User({ email: args.data.email });
+    const password = await hashPassword(args.data.password);
+    const user = await prisma.mutation.createUser({
+      data: {
+        ...args.data,
+        password
+      }
+    });
 
-    if (emailTaken) {
-      throw new Error('Email taken');
-    }
-
-    const user = await prisma.mutation.createUser({ data: args.data }, info);
-
-    return user;
+    return {
+      user,
+      token: generateToken(user.id)
+    };
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id });
-
-    if (!userExists) {
-      throw new Error('User not found');
-    }
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
 
     const user = await prisma.mutation.deleteUser(
       {
         where: {
-          id: args.id
+          id: userId
         }
       },
       info
     );
+
     return user;
   },
-  async updateUser(parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id });
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
 
-    if (!userExists) {
-      throw new Error('User not found');
+    if (typeof args.data.password === 'string') {
+      args.data.password = await hashPassword(args.data.password);
     }
 
     return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id
+          id: userId
         },
         data: args.data
       },
       info
     );
   },
-  async createPost(parent, args, { prisma }, info) {
+  async createPost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
     return prisma.mutation.createPost(
       {
         data: {
@@ -53,7 +60,7 @@ const Mutation = {
           published: args.data.published,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           }
         }
@@ -61,7 +68,18 @@ const Mutation = {
       info
     );
   },
-  async deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+
+    if (!postExists) {
+      throw new Error('Unable to delete post');
+    }
     return prisma.mutation.deletePost(
       {
         where: {
@@ -71,7 +89,28 @@ const Mutation = {
       info
     );
   },
-  async updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    const isPublished = prisma.exists.Post({ id: args.id, published: true });
+
+    if (!postExists) {
+      throw new Error('Unable to update post');
+    }
+    if (isPublished && args.data.published === false) {
+      await prisma.mutation.deleteManyComments({
+        where: {
+          post: {
+            id: args.id
+          }
+        }
+      });
+    }
     return prisma.mutation.updatePost(
       {
         where: {
@@ -82,14 +121,22 @@ const Mutation = {
       info
     );
   },
-  async createComment(parent, args, { prisma }, info) {
+  async createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.data.post,
+      published: true
+    });
+    if (!postExists) {
+      throw new Error('Post does not exist');
+    }
     return prisma.mutation.createComment(
       {
         data: {
           text: args.data.text,
           author: {
             connect: {
-              id: args.data.author
+              id: userId
             }
           },
           post: {
@@ -102,7 +149,17 @@ const Mutation = {
       info
     );
   },
-  async deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!commentExists) {
+      throw new Error('Unable to delete comment');
+    }
     return prisma.mutation.deleteComment(
       {
         where: {
@@ -112,7 +169,17 @@ const Mutation = {
       info
     );
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if (!commentExists) {
+      throw new Error('Unable to update comment');
+    }
     return prisma.mutation.updateComment(
       {
         where: {
@@ -122,6 +189,26 @@ const Mutation = {
       },
       info
     );
+  },
+  async login(parent, args, { prisma }, info) {
+    const user = await prisma.query.user({
+      where: {
+        email: args.data.email
+      }
+    });
+    if (!user) {
+      throw new Error('Unable to login');
+    }
+
+    const isMatch = await bcrypt.compare(args.data.password, user.password);
+
+    if (!isMatch) {
+      throw new Error('Unable to login');
+    }
+    return {
+      user,
+      token: generateToken(user.id)
+    };
   }
 };
